services/auth-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
billiard==4.2.2
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-phonenumber-field
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.9.0
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-phonenumber-field==8.3.0
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-yasg==1.21.11
    # via -r requirements.in
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   drf-yasg
    #   kombu
phonenumbers==9.0.16
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.3
    # via drf-yasg
redis==6.4.0
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via kombu
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
